# Tp1-ServiceWeb/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TP1",
        "version": "0.1"
    },
    "paths": {
        "/api/films/{id}/actors": {
            "get": {
                "tags": [
                    "Actors"
                ],
                "summary": "Obtenir la liste des acteurs d'un film",
                "description": "Retourne une liste des acteurs associés à l'ID d'un film donné",
                "operationId": "d62b9260915205f0765620cea8792c4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des acteurs récupérée avec succès"
                    },
                    "404": {
                        "description": "Film introuvable",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "FilmIntrouvable": {
                                        "summary": "Réponse lorsque le film n'est pas trouvé",
                                        "value": {
                                            "error": "Film introuvable",
                                            "message": "Aucun résultat trouvé pour le modèle [Film] avec cet ID"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "ID invalide",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "IDInvalide": {
                                        "summary": "Réponse lorsque l'ID est incorrect",
                                        "value": {
                                            "error": "ID invalide",
                                            "message": "L'ID du film doit être un entier positif."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur de base de données",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "ErreurBaseDeDonnees": {
                                        "summary": "Réponse en cas d'erreur de base de données",
                                        "value": {
                                            "error": "Erreur de base de données",
                                            "message": "SQLSTATE[HY000]: Erreur générale"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/films/{id}/critics": {
            "get": {
                "tags": [
                    "Critics"
                ],
                "summary": "Obtenir les critiques d'un film",
                "description": "Retourne une liste de critiques associées à un film donné",
                "operationId": "7628d1cb411607ea9479f275ab647dfe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des critiques récupérée avec succès"
                    },
                    "404": {
                        "description": "Film introuvable",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "FilmIntrouvable": {
                                        "summary": "Film non trouvé",
                                        "value": {
                                            "error": "Film non trouvé",
                                            "message": "Aucun film trouvé avec cet ID"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "ID invalide",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "IDInvalide": {
                                        "summary": "ID fourni non valide",
                                        "value": {
                                            "error": "ID invalide",
                                            "message": "L'ID du film doit être un entier positif."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Erreur base de données",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "ErreurBaseDeDonnees": {
                                        "summary": "Erreur de base de données",
                                        "value": {
                                            "error": "Erreur base de données",
                                            "message": "SQLSTATE[HY000]: Erreur générale"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "ErreurServeur": {
                                        "summary": "Erreur interne du serveur",
                                        "value": {
                                            "error": "Erreur serveur",
                                            "message": "Une erreur inconnue s'est produite."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/critics/{id}": {
            "delete": {
                "tags": [
                    "Critics"
                ],
                "summary": "Supprimer une critique",
                "description": "Supprime une critique spécifique par son ID",
                "operationId": "4c7bbd907e57760bbbfc560c1ba8e1ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la critique",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Critique supprimée avec succès"
                    },
                    "404": {
                        "description": "Critique non trouvée",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "CritiqueIntrouvable": {
                                        "summary": "Critique non trouvée",
                                        "value": {
                                            "error": "Critique non trouvée",
                                            "message": "Aucune critique trouvée avec cet ID"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "ID invalide",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "IDInvalide": {
                                        "summary": "ID fourni non valide",
                                        "value": {
                                            "error": "ID invalide",
                                            "message": "L'ID de la critique doit être un entier positif."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Erreur base de données",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "ErreurBaseDeDonnees": {
                                        "summary": "Erreur de base de données",
                                        "value": {
                                            "error": "Erreur base de données",
                                            "message": "SQLSTATE[HY000]: Erreur générale"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "ErreurServeur": {
                                        "summary": "Erreur interne du serveur",
                                        "value": {
                                            "error": "Erreur serveur",
                                            "message": "Une erreur inconnue s'est produite."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Actors",
            "description": "Actors"
        },
        {
            "name": "Critics",
            "description": "Critics"
        }
    ]
}